// src/api/generated/schema.d.ts
export interface components {
    schemas: {
// ...
        Feeding: {
            readonly id?: number;
            /** @description Required unless a Timer value is provided. */
            child?: number | null;
            /**
             * Format: date-time
             * @description Required unless a Timer value is provided.
             */
            start?: string;
            /**
             * Format: date-time
             * @description Required unless a Timer value is provided.
             */
            end?: string;
            /** @description May be used in place of the Start, End, and/or Child values. */
            timer?: number | null;
            readonly duration?: string | null;
            /** @enum {string} */
            type: "breast milk" | "formula" | "fortified breast milk" | "solid food";
            /** @enum {string} */
            method: "bottle" | "left breast" | "right breast" | "both breasts" | "parent fed" | "self fed";
            amount?: number | null;
            notes?: string | null;
            tags?: string[];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}